+++ src/chain_probe.h
@@ -1,5 +1,9 @@
 #pragma once
 
+#include <Arduino.h>
+
 void chainProbeSetup();
 void chainProbeUpdate();
-
+size_t chainProbeKeyCount();
+String chainProbeKeyUid(size_t index);
+bool chainProbeKeyConnected(const String& uid);